+++ js/src/main/scala/inrae/semantic_web/strategy/StrategyRequestBuilder.scala
@@ -1,7 +1,8 @@
 package inrae.semantic_web.strategy
 
 import inrae.semantic_web.driver.ComunicaFederatedStrategy
-import inrae.semantic_web.{SWDiscoveryException, StatementConfiguration}
+import inrae.semantic_web.exception._
+import inrae.semantic_web.configuration._
 
 /**
  * Build a strategy to request a set of web sem sources (triple store/file/inline turtle)
