+++ js/src/test/scala/inrae/semantic_web/SWTransactionJsTest.scala
@@ -1,6 +1,7 @@
 package inrae.semantic_web
 import scala.scalajs.js.JSConverters._
 import inrae.data.DataTestFactory
+import inrae.semantic_web.configuration._
 import utest.{TestSuite, Tests, test}
 
 object SWTransactionJsTest extends TestSuite{
