+++ jvm/src/main/scala/inrae/semantic_web/driver/Rdf4jRequestDriver.scala
@@ -1,6 +1,6 @@
 package inrae.semantic_web.driver
 
-import inrae.semantic_web.SWDiscoveryException
+import inrae.semantic_web.exception._
 import inrae.semantic_web.event.{DiscoveryRequestEvent, DiscoveryStateRequestEvent}
 import inrae.semantic_web.sparql.QueryResult
 import org.eclipse.rdf4j.query.resultio.sparqljson.SPARQLResultsJSONWriter
