+++ jvm/src/main/scala/inrae/semantic_web/driver/RequestDriverFactory.scala
@@ -1,7 +1,7 @@
 package inrae.semantic_web.driver
 
-import inrae.semantic_web.ConfigurationObject.Source
-import inrae.semantic_web.SWDiscoveryException
+import inrae.semantic_web.exception._
+import inrae.semantic_web.configuration.ConfigurationObject.Source
 import org.eclipse.rdf4j.repository.sail.SailRepository
 import org.eclipse.rdf4j.repository.{Repository, RepositoryConnection}
 import org.eclipse.rdf4j.rio.RDFFormat
