+++ shared/src/main/scala/inrae/semantic_web/exception/SWDiscoveryException.scala
@@ -1,0 +1,4 @@
+package inrae.semantic_web.exception
+
+final case class SWDiscoveryException(private val message: String = "",
+                                      private val cause: Throwable = None.orNull) extends Exception(message,cause)
