+++ shared/src/test/scala/inrae/semantic_web/Fix83PrefixUnknownTest.scala
@@ -2,6 +2,7 @@ package inrae.semantic_web
 
 import inrae.data.DataTestFactory
 import inrae.semantic_web.rdf.URI
+import inrae.semantic_web.configuration._
 import utest.{TestSuite, Tests, assert, test}
 
 import scala.concurrent.ExecutionContext.Implicits.global
