+++ shared/src/test/scala/inrae/semantic_web/SWDiscoveryDecorationTest.scala
@@ -3,6 +3,7 @@ package inrae.semantic_web
 import inrae.data.DataTestFactory
 import inrae.semantic_web.node.{Node, Root}
 import inrae.semantic_web.rdf.URI
+import inrae.semantic_web.configuration._
 import utest._
 
 import scala.language.postfixOps
