+++ shared/src/test/scala/inrae/semantic_web/SWDiscoveryNodeAddTest.scala
@@ -2,6 +2,7 @@ package inrae.semantic_web
 
 import inrae.data.DataTestFactory
 import inrae.semantic_web.rdf._
+import inrae.semantic_web.configuration._
 import utest._
 
 import scala.language.postfixOps
