+++ shared/src/test/scala/inrae/semantic_web/SWTransactionTest.scala
@@ -1,6 +1,7 @@
 package inrae.semantic_web
 
 import inrae.semantic_web.SWDiscoveryTest.config
+import inrae.semantic_web.configuration._
 import utest.{TestSuite, Tests, test}
 
 object SWTransactionTest extends TestSuite {
